requirements.txt
fastapi==0.95.2
pydantic==1.10.2
uvicorn==0.22.0
upsonic
python-dotenv==1.0.0
mcp-server-sentry 
requests
